experimentation/requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
scipy>=1.10.0
python-dotenv>=1.0.0

# Manim Animation Generator (optional - needed for rendering)
manim>=0.18.0
ffmpeg-python>=0.2.0
